filter/requirements.txt
appnope @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_46iquy4y0r/croot/appnope_1750774774351/work
asttokens @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_11i8cnwuxs/croot/asttokens_1743630449718/work
attrs @ file:///opt/miniconda3/conda-bld/attrs_1762356972456/work
comm @ file:///opt/miniconda3/conda-bld/comm_1763119153703/work
contourpy==1.3.3
cycler==0.12.1
debugpy @ file:///opt/miniconda3/conda-bld/debugpy_1762421636036/work
decorator @ file:///opt/miniconda3/conda-bld/decorator_1757341235959/work
executing @ file:///opt/miniconda3/conda-bld/executing_1757061235776/work
fastjsonschema @ file:///opt/miniconda3/conda-bld/python-fastjsonschema_1763718719073/work
filterpy==1.4.5
fonttools==4.61.1
ipykernel @ file:///opt/miniconda3/conda-bld/ipykernel_1762946072373/work
ipython @ file:///opt/miniconda3/conda-bld/ipython_1762789298678/work
ipython_pygments_lexers @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_25fteymumy/croot/ipython_pygments_lexers_1744753256449/work
jedi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_38ctoinnl0/croot/jedi_1733987402850/work
jsonschema==4.26.0
jsonschema-specifications @ file:///opt/miniconda3/conda-bld/jsonschema-specifications_1762788416887/work
jupyter_client @ file:///Users/ec2-user/croot/jupyter_client_1768399009291/work
jupyter_core @ file:///opt/miniconda3/conda-bld/jupyter_core_1764588284546/work
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline @ file:///opt/miniconda3/conda-bld/matplotlib-inline_1762779171588/work
narwhals==2.17.0
nbformat @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_2cv_qoc1gw/croot/nbformat_1728049423516/work
nest-asyncio @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_310vb5e2a0/croot/nest-asyncio_1708532678212/work
numpy==2.4.2
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079066/work
parso @ file:///opt/miniconda3/conda-bld/parso_1762781703292/work
pexpect @ file:///opt/miniconda3/conda-bld/pexpect_1762535935939/work
pillow==12.1.1
platformdirs @ file:///opt/miniconda3/conda-bld/platformdirs_1762356624685/work
plotly==6.5.2
prompt_toolkit @ file:///opt/miniconda3/conda-bld/prompt-toolkit_1761744881146/work
psutil @ file:///opt/miniconda3/conda-bld/psutil_1761896323688/work
ptyprocess @ file:///opt/miniconda3/conda-bld/ptyprocess_1762424170819/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=3470be7f810474c8a2ecfcd6e02acc6aea8483ab595417fa4e336362a349933e
pure_eval @ file:///opt/miniconda3/conda-bld/pure_eval_1757067067189/work
Pygments @ file:///opt/miniconda3/conda-bld/pygments_1762431577787/work
pyparsing==3.3.2
pyserial==3.5
python-dateutil @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_66ud1l42_h/croot/python-dateutil_1716495741162/work
pyzmq @ file:///opt/miniconda3/conda-bld/pyzmq_1762375616040/work
referencing @ file:///opt/miniconda3/conda-bld/referencing_1762536886664/work
rpds-py @ file:///opt/miniconda3/conda-bld/rpds-py_1762366478996/work
scipy==1.17.1
six @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_03myqm7p6o/croot/six_1744271511946/work
stack_data @ file:///opt/miniconda3/conda-bld/stack_data_1757067039886/work
tornado @ file:///Users/ec2-user/croot/tornado_1765992648472/work
traitlets @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_500m2_1wyk/croot/traitlets_1718227071952/work
typing_extensions @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_dffmcwqo6o/croot/typing_extensions_1756281471323/work
wcwidth @ file:///Users/ec2-user/croot/wcwidth_1767780856263/work
